Yellow Ribbon Program support
Provide the ability to upload Yellow Ribbon Program data provided by EDU
team. A new file upload option: "Yellow Ribbon Program Source" is
provided. General validation and conversion is supported, though some
fields that are only used on re-export of data source will need to be
worked through with the EDU team in more detail to ensure we have
consistent and correct input data.

Yellow Ribbon Program Source records are created for each row in the
input file. The Institution Builder is modified to create a new Yellow
Ribbon Program record for each program that can be associated with an
institution. Noticing that we're not seeing a full match to all programs
in the sample data file, so we may need some sort of cross-reference
check with institutions, though the naive approach to this would require
that source files be uploaded in a specific order, so we're going to
want to avoid that.

The Yellow Ribbon Program records corresponding with an Institution are
exposed as part of the institution profile endpoint. Swagger
documentation in Vets-API will be provided shortly to reference this new
functionality.

```
{
   ...
   "data" : {
      "id" : "1565",
      "attributes" : {
         "facility_code" : "21905928",
         ...
         "yellow_ribbon_programs" : [
            {
               "division_professional_school" : "All",
               "number_of_students" : 1000,
               "degree_level" : "All",
               "contribution_amount" : "10000.0"
            }
         ],
	...
}
```

.