STEM program availability
Provides a `stem_offered` boolean value on an `Institution`. Exposes
this value in the institution index as a new facet (filterable with
the `stem_offered` GET parameter, and to the  institution profile
endpoint.

This value is determined based on two new data files, IpedsCipCode and
StemCipCode. The former contains a reference to programs by CIP code
that are supported by an institution, and the latter contains a set of
CIP codes that are applied to STEM programs. The Institution Builder
will correlate the two new data sources to determine if an institution
offers a STEM program.

The `institutions.csv` export contains a `stem_offered` boolean value.