new gem (#633)
Main PR: https://github.com/department-of-veterans-affairs/gibct-data-service/pull/630
https://github.com/ulfurinn/wongi-engine/
https://ulfurinn.github.io/wongi-engine/

Adding a [Rete algorithm](http://en.wikipedia.org/wiki/Rete_algorithm) engine to be used for applying mapping rules to `caution_flags` row as the values provided in the source csvs are not the values that will be displayed to frontend.  This is being designed in a way that could eventually have a GIDs UI component that allows users to create or modify the mapping rules.