VEBT-1165 Fix EightKeys file download URL (#1310)
* VEBT-1165 Fix EightKeys file download URL

The current URL for the Eight Keys file requires following several redirects before the file is actually delivered. Since curl doesn't follow redirects by default, the wrong data is 'downloaded'. I.e. is the html of a 301 redirect page that is downloaded and treated as a xls file, which of course breaks things. The easy fix is to update the URL to the end of the redirect chain that actually delivers the file.

* fix no_key_api_downloader spec based on newly-implemented EightKey url